orchestrator/pyproject.toml
[project]
name = "orchestrator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "ZANA Contributors" }
]
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "langchain>=1.2.15",
    "langchain-anthropic>=1.4.0",
    "langchain-community>=0.4.1",
    "langgraph>=1.1.6",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
]

[build-system]
requires = ["uv_build>=0.10.4,<0.11.0"]
build-backend = "uv_build"

[tool.uv]
package = false
